only send notifications if there are registered devices

+++ Backend/src/utils/notificationSender.js
@@ -14,48 +14,50 @@ const sendNotification = async ({ notification, users }) => {
       });
     });
 
-    const { results } = await admin.messaging().sendToDevice(
-      fcmTokens,
-      {
-        notification,
-      },
-      {
-        contentAvailable: true,
-        priority: 'high',
-      }
-    );
+    if (fcmTokens.length > 0) {
+      const { results } = await admin.messaging().sendToDevice(
+        fcmTokens,
+        {
+          notification,
+        },
+        {
+          contentAvailable: true,
+          priority: 'high',
+        }
+      );
 
-    const devicesToRemove = [];
-    results.forEach((result, index) => {
-      if (result.error) {
-        devicesToRemove.push(deviceIds[index]);
-      }
-    });
+      const devicesToRemove = [];
+      results.forEach((result, index) => {
+        if (result.error) {
+          devicesToRemove.push(deviceIds[index]);
+        }
+      });
 
-    if (devicesToRemove.length > 0) {
-      const session = await mongoose.startSession();
-      session.startTransaction();
-      try {
-        await Device.deleteMany(
-          {
-            _id: { $in: devicesToRemove },
-          },
-          { session: session }
-        );
+      if (devicesToRemove.length > 0) {
+        const session = await mongoose.startSession();
+        session.startTransaction();
+        try {
+          await Device.deleteMany(
+            {
+              _id: { $in: devicesToRemove },
+            },
+            { session: session }
+          );
 
-        await User.updateMany(
-          { devices: { $not: { $size: 0 } } },
-          {
-            $pull: { devices: { $in: devicesToRemove } },
-          },
-          { session: session }
-        );
+          await User.updateMany(
+            { devices: { $not: { $size: 0 } } },
+            {
+              $pull: { devices: { $in: devicesToRemove } },
+            },
+            { session: session }
+          );
 
-        await session.commitTransaction();
-        session.endSession();
-      } catch (error) {
-        await session.abortTransaction();
-        session.endSession();
+          await session.commitTransaction();
+          session.endSession();
+        } catch (error) {
+          await session.abortTransaction();
+          session.endSession();
+        }
       }
     }
   } catch (error) {
